Refactored the logic for set status from board to status. Added style to status
