+++ src/App.js
@@ -13,23 +13,26 @@ function App() {
 
     const handleReset = () => {
         setSquares(Array(9).fill(null))
-        setStatus("Next player: X")
         setIsCounterX(true)
         setEndGame(false)
     }
 
   return (
         <div className="App flex flex-col w-full py-4 gap-10 items-center justify-center">
-            <Status status={status} />
+            <Status
+                status={status}
+                setStatus={setStatus}
+                squares={squares}
+                isCounterX={isCounterX}
+                setEndGame={setEndGame}
+            />
 
             <Board
                 squares={squares}
                 setSquares={setSquares}
-                setStatus={setStatus}
                 isCounterX={isCounterX}
                 setIsCounterX={setIsCounterX}
                 endGame={endGame}
-                setEndGame={setEndGame}
             />
 
             <ResetButton handleReset={handleReset} />
